[audio_to_spectrogram] Add data_amplitude_plot (not audio-specific version of audio_amplitude_plot)
