--- audio_to_spectrogram/cfg/AudioAmplitudePlot.cfg
@@ -1,13 +1,0 @@
-#! /usr/bin/env python
-
-PACKAGE = 'audio_to_spectrogram'
-
-from dynamic_reconfigure.parameter_generator_catkin import *
-
-
-gen = ParameterGenerator()
-
-gen.add("maximum_amplitude", int_t, 0, "Maximum range of amplitude to plot", 10000, 1, 100000)
-gen.add("window_size", double_t, 0,  "Window size of sound input to plot", 10.0, 0.1, 10000)
-
-exit(gen.generate(PACKAGE, PACKAGE, "AudioAmplitudePlot"))
