+++ audio_to_spectrogram/scripts/data_amplitude_plot.py
@@ -1,0 +1,11 @@
+#!/usr/bin/env python
+
+import rospy
+
+from audio_to_spectrogram import DataAmplitudePlot
+
+
+if __name__ == '__main__':
+    rospy.init_node('data_amplitude_plot')
+    DataAmplitudePlot()
+    rospy.spin()
