+++ audio_to_spectrogram/src/audio_to_spectrogram/__init__.py
@@ -1,5 +1,6 @@
 from audio_to_spectrogram.audio_buffer import AudioBuffer
 from audio_to_spectrogram.compat import check_matplotlib_version
 from audio_to_spectrogram.convert_matplotlib import convert_matplotlib_to_img
+from audio_to_spectrogram.data_amplitude_plot import DataAmplitudePlot
 from audio_to_spectrogram.data_buffer import DataBuffer
 from audio_to_spectrogram.data_to_spectrum import DataToSpectrum
